requirements.txt
pdfplumber>=0.11.0
python-docx>=1.1.0
pdf2docx>=0.5.8
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
